chore: organize packages
