+++ src/timeline-service/src/main/java/org/palette/easeltimelineservice/config/KafkaConsumerConfig.java
@@ -1,7 +1,8 @@
-package org.palette.easeltimelineservice;
+package org.palette.easeltimelineservice.config;
 
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.common.serialization.StringDeserializer;
+import org.palette.easeltimelineservice.dto.PaintCreatedEvent;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
