+++ src/timeline-service/src/main/java/org/palette/easeltimelineservice/dto/Entities.java
@@ -1,4 +1,4 @@
-package org.palette.easeltimelineservice;
+package org.palette.easeltimelineservice.dto;
 
 import java.util.List;
