+++ src/timeline-service/src/main/java/org/palette/easeltimelineservice/dto/MediaResponse.java
@@ -1,4 +1,4 @@
-package org.palette.easeltimelineservice;
+package org.palette.easeltimelineservice.dto;
 
 public record MediaResponse(
         String type,
