+++ src/timeline-service/src/main/java/org/palette/easeltimelineservice/dto/MentionResponse.java
@@ -1,4 +1,4 @@
-package org.palette.easeltimelineservice;
+package org.palette.easeltimelineservice.dto;
 
 public record MentionResponse(
         Integer start,
