+++ src/timeline-service/src/main/java/org/palette/easeltimelineservice/dto/UserResponse.java
@@ -1,4 +1,4 @@
-package org.palette.easeltimelineservice;
+package org.palette.easeltimelineservice.dto;
 
 public record UserResponse(
         Long id,
