+++ src/timeline-service/src/main/java/org/palette/easeltimelineservice/service/PaintCacheService.java
@@ -1,6 +1,7 @@
-package org.palette.easeltimelineservice;
+package org.palette.easeltimelineservice.service;
 
 import lombok.RequiredArgsConstructor;
+import org.palette.easeltimelineservice.dto.PaintResponse;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +11,7 @@ public class PaintCacheService {
 
     private final RedisTemplate<String, Object> redisTemplate;
 
-    public void cachePaint(Long paintId, Paint paint) {
-        redisTemplate.opsForValue().set("paint:" + paintId, paint);
+    public void cachePaint(Long paintId, PaintResponse paintResponse) {
+        redisTemplate.opsForValue().set("paint:" + paintId, paintResponse);
     }
 }
