+++ src/timeline-service/src/main/java/org/palette/easeltimelineservice/usecase/TimelineUsecase.java
@@ -1,6 +1,10 @@
-package org.palette.easeltimelineservice;
+package org.palette.easeltimelineservice.usecase;
 
 import lombok.RequiredArgsConstructor;
+import org.palette.easeltimelineservice.service.FollowerPaintMapService;
+import org.palette.easeltimelineservice.service.PaintCacheService;
+import org.palette.easeltimelineservice.external.grpc.GrpcSocialClient;
+import org.palette.easeltimelineservice.dto.PaintCreatedEvent;
 import org.palette.grpc.GFollowerIdsResponse;
 import org.palette.grpc.GSocialServiceGrpc;
 import org.springframework.stereotype.Service;
@@ -15,7 +19,7 @@ public class TimelineUsecase extends GSocialServiceGrpc.GSocialServiceImplBase {
 
     public void handlePaintCreatedEvent(PaintCreatedEvent paintCreatedEvent) {
         final GFollowerIdsResponse followerIds = gRPCSocialClient.getFollowerIds(paintCreatedEvent.userId());
-        paintCacheService.cachePaint(paintCreatedEvent.paintId(), paintCreatedEvent.paint());
+        paintCacheService.cachePaint(paintCreatedEvent.paintId(), paintCreatedEvent.paintResponse());
         followerPaintMapService.addPaintToFollowersTimeline(followerIds.getFollowerIdsList(), paintCreatedEvent.paintId());
     }
 }
