+++ src/timeline-service/src/test/java/org/palette/easeltimelineservice/service/FollowerPaintMapServiceTest.java
@@ -1,8 +1,9 @@
-package org.palette.easeltimelineservice;
+package org.palette.easeltimelineservice.service;
 
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.palette.easeltimelineservice.service.FollowerPaintMapService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.redis.core.RedisTemplate;
